requirements.txt
pandas-ta
python-dotenv
requests
ib_insync